requirements.txt
kafka-python
six
pymongo